Functions

ParticleInCell.compute_b_from_e! — Method

compute_b_from_e!( field_out, maxwell_solver, delta_t, field_in)

Compute Bz from Ey using strong 1D Faraday equation for spline coefficients

\[B_z^{new}(x_j) = B_z^{old}(x_j) - \frac{\Delta t}{\Delta x} (E_y(x_j) - E_y(x_{j-1})\]

ParticleInCell.compute_e_from_b! — Method

compute_e_from_b!(field_out, maxwell_solver, delta_t, field_in)

compute Ey from Bz using weak Ampere formulation

ParticleInCell.compute_e_from_j! — Method

compute_e_from_j!(e, maxwell_solver, current, component)

Compute $E_i$ from $j_i$ integrated over the time interval using weak Ampere formulation

ParticleInCell.compute_rhs_from_function! — Method

computerhsfromfunction(self, func, degree, coefsdofs)

Compute the FEM right-hand-side for a given function f and periodic splines of given degree.

Its components are $\int f N_i dx$ where $N_i$ is the B-spline starting at $x_i$.

ParticleInCell.eval_uniform_periodic_spline_curve — Method

eval_uniform_periodic_spline_curve( degree, scoef )

Evaluate uniform periodic spline curve defined by coefficients scoef at knots (which are the grid points)

ParticleInCell.get_cell_and_offset — Method

get_cell_and_offset( mesh, x )

Get cell and offset

We compute the cell indices where the particle is and its relative normalized position inside the cell

ParticleInCell.get_x — Method

get_x( mesh, i )

Get position

ParticleInCell.inner_product — Method

inner_product( maxwell_solver, coefs1_dofs, coefs2_dofs, degree )

maxwell_solver : Maxwell solver object

coefs1_dofs : Coefficient for each DoF

coefs2_dofs : Coefficient for each DoF

`degree : Specify the degree of the basis functions

return squared L2 norm

ParticleInCell.l2norm_squared — Method

l2norm_squared(maxwell_solver, coefs_dofs, degree)

Compute square of the L2norm

ParticleInCell.l2norm_squared2 — Method

l2norm_squared(maxwell_solver, coefs_dofs, degree)

Compute square of the L2norm

ParticleInCell.l2projection! — Method

l2projection!(coefs_dofs, maxwell, func, degree)

Compute the L2 projection of a given function f on periodic splines of given degree

ParticleInCell.solve! — Method

Compute electric field from charge density

ParticleInCell.uniform_bsplines_eval_basis — Method

uniform_bsplines_eval_basis( spline_degree, normalized_offset, bspl )

UNIFORM B-SPLINE FUNCTIONS

Evaluate all non vanishing uniform B-Splines in unit cell.

Returns an array with the values of the b-splines of the requested degree, evaluated at a given cell offset. The cell size is normalized between 0 and 1, thus the offset given must be a number between 0 and 1.

Output:

\[bspl(1:d+1)= B_d(-(d+1)/2+d+x),...,B_d(-(d+1)/2+x)\]

with $d$=spline_degree and $x$=normalized_offset where $B_d=B_{d-1}*B_0$ and $B_0=1_[-1/2,1/2]$ and * is convolution the following FORTRAN code can be used for comparison with deboor

do i=-d,d+1
    t(i+d+1)=real(i,8)
end do
call bsplvb(t,d+1,1,normalized_offset,d+1,out)

We also have the property (from the symmetry of the B-spline)

\[out[1:d+1]= B_d(-(d+1)/2+xx),...,B_d(-(d+1)/2+d+xx),..., \]

where $xx=1-$ normalized_offset

ParticleInCell.update_positions! — Method

update particle position xp

ParticleInCell.update_velocities! — Method

update particle velocities vp