Functions

ParticleInCell.add_charge! — Method

add_charge(pm, x_position, y_position wp, ρ_dofs)

Add charge of single particle

position : Particle position

wp : Particle weight times charge

ρ_dofs : spline coefficient of accumulated density

ParticleInCell.add_charge! — Method

add_charge!( rho, p, position, marker_charge)

Add charge of one particle

p : kernel smoother object

position : Position of the particle

marker_charge : Particle weights time charge

rho_dofs : Coefficient vector of the charge distribution

ParticleInCell.add_charge! — Method

add_charge!(pic, position, marker_charge)

Add charge from one particle

self : Pic Poisson solver object

position : Position of the particle

marker_charge : Particle weight times charge

ParticleInCell.add_charge_pp! — Method

add_charge_single_spline_pp_2d(pm, position, wp, ρ_dofs)

Add charge of single particle

Information about the 2d mesh

delta_x(2) : Value of grid spacing along both directions.

domain(2,2) : Definition of the domain: domain(1,1) = x1min, domain(2,1) = x2min, domain(1,2) = x1max, domain(2,2) = x2max

Information about the particles

no_particles : Number of particles of underlying PIC method (processor local)

n_span : Number of intervals where spline non zero (degree + 1)

scaling

position : Particle position

wp : Particle weight times charge

ρ_dofs : spline coefficient of accumulated density

ParticleInCell.add_charge_pp! — Method

add_charge_pp!(rho_dofs, p, position, marker_charge)

Add charge of one particle

p : kernel smoother object

position : Position of the particle

marker_charge : Particle weights time charge

rho_dofs : Coefficient vector of the charge distribution

ParticleInCell.add_current_update_v! — Method

add_current_update_v!( j_dofs, p, 
                       position_old, position_new, 
                       marker_charge, qoverm, 
                       bfield_dofs, vi)

Add current for one particle and update v (according to $H_{p1}$ part in Hamiltonian splitting)

Read out particle position and velocity

Compute index_old, the index of the last DoF on the grid the

particle contributes to, and r_old, its position (normalized to cell size one).

ParticleInCell.add_current_update_v_pp! — Method

add_current_update_v_pp!( j_dofs, p, position_old, position_new, 
                          marker_charge, qoverm, bfield_dofs, vi)

Add current for one particle and update v (according to H_p1 part in Hamiltonian splitting)

ParticleInCell.b_to_pp — Method

b_to_pp( SplinePP, ncells, b_coeffs)

Convert 1d spline in B form to spline in pp form with periodic boundary conditions

ParticleInCell.b_to_pp_1d_cell! — Method

b_to_pp_1d_cell( self, b_coeffs, pp_coeffs )

Convert 1d spline in B form in a cell to spline in pp form with periodic boundary conditions

spline : arbitrary degree 1d spline

b_coeffs(self%degree+1) : coefficients of spline in B-form

pp_coeffs(self%degree+1) : coefficients of spline in pp-form

ParticleInCell.b_to_pp_2d! — Method

b_to_pp_2d!( pp, spl1, spl2, b)

Convert 2d spline in B form to spline in pp form

n_cells(2) : number of gridcells

b_coeffs : coefficients of spline in B-form

pp_coeffs : coefficients of spline in pp-form

ParticleInCell.b_to_pp_2d_cell! — Method

b_to_pp_2d_cell(spline1, spline2, b_coeffs, pp_coeffs, i, j)

Convert 2d spline in B form in a cell to spline in pp form with periodic boundary conditions

spline1 : arbitrary degree 1d spline

spline2 : arbitrary degree 1d spline

n_cells(2) : number of gridcells

bcoeffs(ncells(1)*n_cells(2)) : coefficients of spline in B-form

ppcoeffs((spline1.degree+1)*(spline2.degree+1),ncells(1)*n_cells(2)) : coefficients of spline in pp-form

ParticleInCell.compute_b_from_e! — Method

compute_b_from_e!( field_out, maxwell_solver, delta_t, field_in)

Compute Bz from Ey using strong 1D Faraday equation for spline coefficients

\[B_z^{new}(x_j) = B_z^{old}(x_j) - \frac{\Delta t}{\Delta x} (E_y(x_j) - E_y(x_{j-1})\]

ParticleInCell.compute_e_from_b! — Method

compute_e_from_b!(field_out, maxwell_solver, delta_t, field_in)

compute Ey from Bz using weak Ampere formulation

ParticleInCell.compute_e_from_j! — Method

compute_e_from_j!(e, maxwell_solver, current, component)

Compute $E_i$ from $j_i$ integrated over the time interval using weak Ampere formulation

ParticleInCell.compute_field_energy — Method

compute_field_energy

Compute the squared l2 norm of electric field

ParticleInCell.compute_rhs_from_function! — Method

computerhsfromfunction(self, func, degree, coefsdofs)

Compute the FEM right-hand-side for a given function f and periodic splines of given degree.

Its components are $\int f N_i dx$ where $N_i$ is the B-spline starting at $x_i$.

ParticleInCell.compute_shape_factor — Method

computeshapefactor(pm, xp, yp)

Helper function computing shape factor

pm : kernel smoother object

ParticleInCell.eval_uniform_periodic_spline_curve — Method

eval_uniform_periodic_spline_curve( degree, scoef )

Evaluate uniform periodic spline curve defined by coefficients scoef at knots (which are the grid points)

ParticleInCell.evaluate — Method

evaluate_field_single_spline_2d(pm, position, field_dofs)

position(pm.dim) : Position where to evaluate

fielddofs(pm.ndofs) : Degrees of freedom in kernel representation.

field_value : Value of the field

Evaluate field with given dofs at position

ParticleInCell.evaluate — Method

evaluate(p, position, field_dofs)

Evaluate field at position

p : Kernel smoother object

position : Position of the particle

field_dofs : Coefficient vector for the field DoFs

field_value : Value(s) of the electric fields at given position

ParticleInCell.evaluate_fields — Method

evaluate_field_single_2d(pic, position, components, func_value)

Evaluate components of the electric field as one position

self : PIC Poisson solver object

components

position : Position of the particle

func_value

ParticleInCell.evaluate_multiple — Method

evaluate_multiple(pm, position, components, field_dofs)

Evaluate multiple fields at position  position

position(pm%dim) : Position where to evaluate

components(:) : Components of the field that shall be evaluated

field_dofs(:,:) : Degrees of freedom in kernel representation.

field_value(:) : Value of the field

ParticleInCell.evaluate_pp — Method

evaluate_pp(p, position, field_dofs_pp)

Evaluate field at position using horner scheme

p : Kernel smoother object

position : Position of the particle

field_dofs_pp : Degrees of freedom in kernel representation.

field_value : Value(s) of the electric fields at given position

ParticleInCell.evaluate_pp — Method

evaluate_pp(pm, position, field_dofs_pp)

Evaluate field at position using horner scheme

pm : kernel smoother object

position : Position where to evaluate

fielddofspp(:,:) : Degrees of freedom in kernel representation.

field_value : Value of the field

ParticleInCell.evaluate_rho! — Method

evaluate_rho!(pic, position)

Evaluate charge density at rho at one position

self : Pic Poisson solver object

position : Position of the particle

func_value : Value of rho at given position

ParticleInCell.evaluate_rho — Method

evaluate_phi!(pic, position)

Evaluate potential at one position

self : Pic Poisson solver object

position : Position of the particle

func_value : Value of phi at given position

ParticleInCell.get_cell_and_offset — Method

get_cell_and_offset( mesh, x )

Get cell and offset

We compute the cell indices where the particle is and its relative normalized position inside the cell

ParticleInCell.get_charge — Method

get_charge( p, i; i_wi=1)

Get charge of ith particle of p (q * particle_weight)

ParticleInCell.get_mass — Method

get_mass( p, i; i_wi=1)

Get mass of ith particle of p (m * particle_weight)

ParticleInCell.get_v — Method

get_v( p, i )

Get velocity of ith particle of p

ParticleInCell.get_weight — Method

get_weight( p, i)

Get ith particle weights of group p

ParticleInCell.get_x — Method

get_x( mesh, i )

Get position

ParticleInCell.get_x — Method

get_x( p, i )

Get position of ith particle of p

ParticleInCell.horner_1d — Method

horner_1d(degree, pp_coeffs, x, index)

Perform a 1d Horner schema on the pp_coeffs at index

ParticleInCell.horner_2d — Method

horner_2d(degrees, pp_coeffs, position, indices, ncells)

Perform a 2d hornerschema on the pp_coeffs at the indices

degree : degree of the spline

pp_coeffs : coefficients of spline in pp-form

position(2) : point at which we evaluate our spline

indices(2) : indices of cell in which is x

ncells(2) : number of gridcells

res : value of the splinefunction at position

ParticleInCell.horner_m_2d! — Method

horner_m_2d!(val, spl1, spl2, degree, x)

Perform two times a 1d hornerschema on the poly_coeffs

val : array of values

degree : degree of the spline

x : point at which we evaluate our spline

ParticleInCell.horner_primitive_1d — Method

horner_primitive_1d(val, degree, pp_coeffs, x)

Perform a 1d Horner schema on the pp_coeffs evaluate at x

ParticleInCell.index_1dto2d_column_major — Method

index_1dto2d_column_major(pm, index1d)

Self function computes the index of a 1D array that stores 2D data in column major ordering. It also takes periodic boundary conditions into account.

index1d_1 !< indice along x (start counting with zero).

index1d_2 !< indice along y (start counting with zero).

index2d !< Corresponding index in 1d array representing 2d data (start counting with one).

ParticleInCell.inner_product — Method

inner_product( maxwell_solver, coefs1_dofs, coefs2_dofs, degree )

maxwell_solver : Maxwell solver object

coefs1_dofs : Coefficient for each DoF

coefs2_dofs : Coefficient for each DoF

`degree : Specify the degree of the basis functions

return squared L2 norm

ParticleInCell.l2norm_squared — Method

l2norm_squared(maxwell_solver, coefs_dofs, degree)

Compute square of the L2norm

ParticleInCell.l2norm_squared2 — Method

l2norm_squared(maxwell_solver, coefs_dofs, degree)

Compute square of the L2norm

ParticleInCell.l2projection! — Method

l2projection!(coefs_dofs, maxwell, func, degree)

Compute the L2 projection of a given function f on periodic splines of given degree

ParticleInCell.operator_t! — Method

operator_t(split, dt)

Push x

split :: time splitting object

dt :: time step

ParticleInCell.sample! — Method

sample!(d, pg)

Sampling from a probability distribution to initialize a Landau damping in 1D2V space.

\[f_0(x,v,t) = \frac{n_0}{2π v_{th}^2} ( 1 + \alpha cos(k_x x)) exp( - \frac{v_x^2+v_y^2}{2 v_{th}^2})\]

The newton function solves the equation $P(x)-r=0$ with Newton’s method

\[x^{n+1} = x^n – (P(x)-(2\pi r / k)/f(x) \]

with

\[P(x) = \int_0^x (1 + \alpha cos(k_x y)) dy = x + \frac{\alpha}{k_x} sin(k_x x)\]

ParticleInCell.sample! — Method

sample!(d, pg)

Sampling from a probability distribution to initialize a Landau damping in 2D2V space.

\[f_0(x,v,t) = \frac{n_0}{2π v_{th}^2} ( 1 + \alpha cos(k_x x)) exp( - \frac{v_x^2+v_y^2}{2 v_{th}^2})\]

ParticleInCell.set_v! — Method

set_v!( p, i, v)

Set velocity of ith particle of p to v

ParticleInCell.set_v! — Method

set_v!( p, i, v)

Set velocity of ith particle of p to v

ParticleInCell.set_w! — Method

set_w!( p, i, w)

Set weights of ith particle of p to w

ParticleInCell.set_w! — Method

set_w!( p, i, w)

Set weights of particle @ i

ParticleInCell.set_x! — Method

set_x!( p, i, x )

Set position of ith particle of p to x

ParticleInCell.set_x! — Method

set_x!( p, i, x)

Set position of ith particle of p to x

Note

if x is a scalar value, only the first x dimension will be set.

ParticleInCell.solve! — Method

Compute electric field from charge density

ParticleInCell.solve! — Method

solve!( ex, ey, poisson, rho )

solves Poisson equation to compute electric fields

\[E(x,y) = -\nabla \phi(x,y) \\ -\Delta \phi(x,y) = \rho(x,y)\]

ParticleInCell.solve! — Method

solve!( poisson, phi, rho )

computes phi from rho

\[-\Delta phi(x,y) = rho(x,y)\]

ParticleInCell.solve! — Method

solve!( pic )

Solve for phi and fields

poisson : Pic Poisson solver object

ParticleInCell.solve_fields! — Method

solve_fields!( pic )

Solve efields from rho

ParticleInCell.solve_fields! — Method

Solve Poisson's equation for the electric field

ParticleInCell.solve_phi! — Method

solve_phi!( pic )

Solve for potential

ParticleInCell.strang_splitting! — Method

Strang splitting

split :: time splitting object

dt :: time step

ParticleInCell.uniform_bsplines_eval_basis — Method

uniform_bsplines_eval_basis( spline_degree, normalized_offset, bspl )

UNIFORM B-SPLINE FUNCTIONS

Evaluate all non vanishing uniform B-Splines in unit cell.

Returns an array with the values of the b-splines of the requested degree, evaluated at a given cell offset. The cell size is normalized between 0 and 1, thus the offset given must be a number between 0 and 1.

Output:

\[bspl(1:d+1)= B_d(-(d+1)/2+d+x),...,B_d(-(d+1)/2+x)\]

with $d$=spline_degree and $x$=normalized_offset where $B_d=B_{d-1}*B_0$ and $B_0=1_[-1/2,1/2]$ and * is convolution the following FORTRAN code can be used for comparison with deboor

do i=-d,d+1
    t(i+d+1)=real(i,8)
end do
call bsplvb(t,d+1,1,normalized_offset,d+1,out)

We also have the property (from the symmetry of the B-spline)

\[out[1:d+1]= B_d(-(d+1)/2+xx),...,B_d(-(d+1)/2+d+xx),..., \]

where $xx=1-$ normalized_offset

ParticleInCell.update_jv! — Method

update_jv!(j_dofs, p, 
           lower, upper, index, marker_charge, 
           qoverm, sign, vi, bfield_dofs)

Helper function for add_current_update_v.

ParticleInCell.update_jv_pp! — Method

update_jv_pp!( j_dofs, p, lower, upper, index, marker_charge, 
               qoverm, vi, bfield_dofs)

Helper function for add_current_update_v.

ParticleInCell.update_positions! — Method

update particle position xp

ParticleInCell.update_velocities! — Method

update particle velocities vp