Functions

ParticleInCell.add_charge! — Method

add_charge!( rho, p, position, marker_charge)

Add charge of one particle

p : kernel smoother object

position : Position of the particle

marker_charge : Particle weights time charge

rho_dofs : Coefficient vector of the charge distribution

ParticleInCell.add_charge_pp! — Method

add_charge_pp!(rho_dofs, p, position, marker_charge)

Add charge of one particle

p : kernel smoother object

position : Position of the particle

marker_charge : Particle weights time charge

rho_dofs : Coefficient vector of the charge distribution

ParticleInCell.add_current_update_v! — Method

add_current_update_v!( j_dofs, p, 
                       position_old, position_new, 
                       marker_charge, qoverm, 
                       bfield_dofs, vi)

Add current for one particle and update v (according to $H_{p1}$ part in Hamiltonian splitting)

Read out particle position and velocity

Compute index_old, the index of the last DoF on the grid the

particle contributes to, and r_old, its position (normalized to cell size one).

ParticleInCell.add_current_update_v_pp! — Method

add_current_update_v_pp!( j_dofs, p, position_old, position_new, 
                          marker_charge, qoverm, bfield_dofs, vi)

Add current for one particle and update v (according to H_p1 part in Hamiltonian splitting)

ParticleInCell.b_to_pp — Method

b_to_pp( SplinePP, ncells, b_coeffs)

Convert 1d spline in B form to spline in pp form with periodic boundary conditions

ParticleInCell.compute_b_from_e! — Method

compute_b_from_e!( field_out, maxwell_solver, delta_t, field_in)

Compute Bz from Ey using strong 1D Faraday equation for spline coefficients

\[B_z^{new}(x_j) = B_z^{old}(x_j) - \frac{\Delta t}{\Delta x} (E_y(x_j) - E_y(x_{j-1})\]

ParticleInCell.compute_e_from_b! — Method

compute_e_from_b!(field_out, maxwell_solver, delta_t, field_in)

compute Ey from Bz using weak Ampere formulation

ParticleInCell.compute_e_from_j! — Method

compute_e_from_j!(e, maxwell_solver, current, component)

Compute $E_i$ from $j_i$ integrated over the time interval using weak Ampere formulation

ParticleInCell.compute_rhs_from_function! — Method

computerhsfromfunction(self, func, degree, coefsdofs)

Compute the FEM right-hand-side for a given function f and periodic splines of given degree.

Its components are $\int f N_i dx$ where $N_i$ is the B-spline starting at $x_i$.

ParticleInCell.eval_uniform_periodic_spline_curve — Method

eval_uniform_periodic_spline_curve( degree, scoef )

Evaluate uniform periodic spline curve defined by coefficients scoef at knots (which are the grid points)

ParticleInCell.evaluate — Method

evaluate(p, position, field_dofs)

Evaluate field at position

p : Kernel smoother object

position : Position of the particle

field_dofs : Coefficient vector for the field DoFs

field_value : Value(s) of the electric fields at given position

ParticleInCell.evaluate_pp — Method

evaluate_pp(p, position, field_dofs_pp)

Evaluate field at position using horner scheme

p : Kernel smoother object

position : Position of the particle

field_dofs_pp : Degrees of freedom in kernel representation.

field_value : Value(s) of the electric fields at given position

ParticleInCell.get_cell_and_offset — Method

get_cell_and_offset( mesh, x )

Get cell and offset

We compute the cell indices where the particle is and its relative normalized position inside the cell

ParticleInCell.get_charge — Method

get_charge( p, i; i_wi=1)

Get charge of ith particle of p (q * particle_weight)

ParticleInCell.get_mass — Method

get_mass( p, i; i_wi=1)

Get mass of ith particle of p (m * particle_weight)

ParticleInCell.get_spin — Method

get_spin( p, i, j)

Get the jth weight of the ith particle weights of group p

ParticleInCell.get_v — Method

get_v( p, i )

Get velocity of ith particle of p

ParticleInCell.get_weights — Method

get_weights( p, i)

Get ith particle weights of group p

ParticleInCell.get_x — Method

get_x( mesh, i )

Get position

ParticleInCell.get_x — Method

get_x( p, i )

Get position of ith particle of p

ParticleInCell.horner_1d — Method

horner_1d(degree, pp_coeffs, x, index)

Perform a 1d Horner schema on the pp_coeffs at index

ParticleInCell.horner_primitive_1d — Method

horner_primitive_1d(val, degree, pp_coeffs, x)

Perform a 1d Horner schema on the pp_coeffs evaluate at x

ParticleInCell.inner_product — Method

inner_product( maxwell_solver, coefs1_dofs, coefs2_dofs, degree )

maxwell_solver : Maxwell solver object

coefs1_dofs : Coefficient for each DoF

coefs2_dofs : Coefficient for each DoF

`degree : Specify the degree of the basis functions

return squared L2 norm

ParticleInCell.l2norm_squared — Method

l2norm_squared(maxwell_solver, coefs_dofs, degree)

Compute square of the L2norm

ParticleInCell.l2norm_squared2 — Method

l2norm_squared(maxwell_solver, coefs_dofs, degree)

Compute square of the L2norm

ParticleInCell.l2projection! — Method

l2projection!(coefs_dofs, maxwell, func, degree)

Compute the L2 projection of a given function f on periodic splines of given degree

ParticleInCell.set_spin — Method

set_spin( p, i, j)

Set the jth weight of the ith particle weights of group p

ParticleInCell.set_v — Method

set_v( p, i, v)

Set velocity of ith particle of p to v

ParticleInCell.set_v — Method

set_v( p, i, v)

Set velocity of ith particle of p to v

ParticleInCell.set_weights — Method

set_weights( p, i, w)

Set weights of ith particle of p to w

ParticleInCell.set_weights — Method

set_weights( p, i, w)

Set weights of particle @ i

ParticleInCell.set_x — Method

set_x( p, i, x )

Set position of ith particle of p to x

ParticleInCell.set_x — Method

set_x( p, i, x)

Set position of ith particle of p to x

Note

if x is a scalar value, only the first x dimension will be set.

ParticleInCell.solve! — Method

Compute electric field from charge density

ParticleInCell.spline_pp_b_to_pp_2d_periodic — Method

Convert 2d spline in B form to spline in pp form with periodic boundary conditions. This is a special case of the procedure splineppbtopp_2d for the double periodic case to avoid the select case statements

self : arbitrary degree 2d spline

n_cells : number of gridcells

bcoeffs(ncells(1)*n_cells(2)) : coefficients of spline in B-form

ppcoeffs((self%spline1%degree+1)*(self%spline2%degree+1),ncells(1)*n_cells(2)) : coefficients of spline in pp-form

ParticleInCell.uniform_bsplines_eval_basis — Method

uniform_bsplines_eval_basis( spline_degree, normalized_offset, bspl )

UNIFORM B-SPLINE FUNCTIONS

Evaluate all non vanishing uniform B-Splines in unit cell.

Returns an array with the values of the b-splines of the requested degree, evaluated at a given cell offset. The cell size is normalized between 0 and 1, thus the offset given must be a number between 0 and 1.

Output:

\[bspl(1:d+1)= B_d(-(d+1)/2+d+x),...,B_d(-(d+1)/2+x)\]

with $d$=spline_degree and $x$=normalized_offset where $B_d=B_{d-1}*B_0$ and $B_0=1_[-1/2,1/2]$ and * is convolution the following FORTRAN code can be used for comparison with deboor

do i=-d,d+1
    t(i+d+1)=real(i,8)
end do
call bsplvb(t,d+1,1,normalized_offset,d+1,out)

We also have the property (from the symmetry of the B-spline)

\[out[1:d+1]= B_d(-(d+1)/2+xx),...,B_d(-(d+1)/2+d+xx),..., \]

where $xx=1-$ normalized_offset

ParticleInCell.update_jv! — Method

update_jv!(j_dofs, p, 
           lower, upper, index, marker_charge, 
           qoverm, sign, vi, bfield_dofs)

Helper function for add_current_update_v.

ParticleInCell.update_jv_pp! — Method

update_jv_pp!( j_dofs, p, lower, upper, index, marker_charge, 
               qoverm, vi, bfield_dofs)

Helper function for add_current_update_v.

ParticleInCell.update_positions! — Method

update particle position xp

ParticleInCell.update_velocities! — Method

update particle velocities vp