Types

ParticleInCell.LandauDamping — Type

Landau( α, kx)

Test structure to initialize a particles distribtion for Landau damping test case in 1D1V and 1D2V

ParticleInCell.OneDPoisson — Type

OneDPoisson

Derived type to solve the Poisson equation on 1d regular cartesian mesh with periodic boundary conditions

kx : wave number in x

nc_x : cells number in x

dx : x step size

rht : fft(rho)

ParticleInCell.OneDPoissonPIC — Type

OneDPoissonPIC( poisson, kernel )

kernel : Kernel smoother taking care of charge deposition and field evaluation

poisson : Poisson solver

rho : Coefficients of expansion of rho (MPI global version)

efield : Coefficients of expansion of electric field

phi : Coefficients of expansion of potential

ParticleInCell.OneDSplittingOperator — Type

Operator splitting type for 2d2v Vlasov-Poisson

pic :: PIC poisson solver

pg :: Particle group

ParticleInCell.SplittingOperator — Type

Operator splitting type for 2d2v Vlasov-Poisson

pic :: PIC poisson solver

pg :: Particle group

ParticleInCell.TwoDPoissonPIC — Type

TwoDPoissonPIC( poisson, kernel )

kernel : Kernel smoother taking care of charge deposition and field evaluation

poisson : Poisson solver

rho_dofs : Coefficients of expansion of rho (MPI global version)

efield_dofs : Coefficients of expansion of electric field

phi_dofs : Coefficients of expansion of potential

rho2d : 2d version of rho_dofs to adjust to field solver format

efield : 2d version of efield_dofs to adjust to field solver format

phi2d : 2d version of phi_dofs to adjust to field solver format

ParticleInCell.TwoDPoissonPeriodic — Type

TwoDPoissonPeriodic

Derived type to solve the Poisson equation on 2d regular cartesian mesh with periodic boundary conditions on both sides

kx : wave number in x

ky : wave number in y

k2 : $k_x^2 + k_y^2$

nc_x : cells number in x

nc_y : cells number in y

dx : x step size

dy : y step size

rht : fft(rho)