Types

ParticleInCell.Maxwell1DFEM — Type

maxwell_solver = MaxwellFEM1D( mesh, degree )

1D Maxwell spline finite element solver on a periodic grid

Lx : length of Periodic domain

delta_x : cell size

n_dofs : number of cells (and grid points)

s_deg_0 : spline degree 0-forms

s_deg_1 : spline degree 1-forms

mass_0 : coefficients of 0-form mass matrix

mass_1 : coefficients of 1-form mass matrix

eig_mass0 : eigenvalues of circulant 0-form mass matrix

eig_mass1 : eigenvalues of circulant 1-form mass matrix

eig_weak_ampere : eigenvalues of circulant update matrix for Ampere

eig_weak_poisson : eigenvalues of circulant update matrix for Poisson

plan_fw : fft plan (forward)

plan_bw : fft plan (backward)

ParticleInCell.OneDGrid — Type

TwoDGrid( xmin, xmax, nx )

Simple structure to store mesh data from 1 to 3 dimensions

ParticleInCell.PICPoisson2D — Type

PICPoisson2D( poisson, kernel )

kernel : Kernel smoother taking care of charge deposition and field evaluation

poisson : Poisson solver

rho_dofs : Coefficients of expansion of rho (MPI global version)

efield_dofs : Coefficients of expansion of electric field

phi_dofs : Coefficients of expansion of potential

rho2d : 2d version of rho_dofs to adjust to field solver format

efield : 2d version of efield_dofs to adjust to field solver format

phi2d : 2d version of phi_dofs to adjust to field solver format

ParticleInCell.ParticleGroup — Type

ParticleGroup{D,V}( n_particles, charge, mass, q, weights)

n_particles : number of particles

charge : charge of the particle species

mass : mass of the particle species

n_weights : number of differents weights

ParticleInCell.ParticleMeshCoupling1D — Type

ParticleMeshCoupling1D( mesh, no_particles, spline_degree, 
                      smoothing_type )

Kernel smoother with splines of arbitrary degree placed on a uniform mesh. Spline with index i starts at point i

delta_x : Value of grid spacing along both directions.

n_grid : Array containing number ofpoints along each direction

no_particles : Number of particles of underlying PIC method

spline_degree : Degree of smoothing kernel spline

n_span : Number of intervals where spline non zero (spline_degree + 1)

scaling : Scaling factor depending on whether :galerkin or :collocation

n_quad_points : Number of quadrature points

spline_val: scratch data for spline evaluation

spline_val_more : more scratch data for spline evaluation

quad_x, quad_w : quadrature weights and points

Note

Only 1D version is implemented for now

ParticleInCell.ParticleMeshCoupling2D — Type

ParticleMeshCoupling2D( pg, grid, degree)

n_grid(2) : no. of spline coefficients

domain(2,2) : lower and upper bounds of the domain

no_particles : no. of particles

degree : Degree of smoothing kernel spline

smoothing_type : Define if Galerkin or collocation smoothing for right scaling in accumulation routines

ParticleInCell.Poisson2DPeriodic — Type

Poisson2DPeriodic

Derived type to solve the Poisson equation on 2d regular cartesian mesh with periodic boundary conditions on both sides

kx : wave number in x

ky : wave number in y

k2 : $k_x^2 + k_y^2$

nc_x : cells number in x

nc_y : cells number in y

dx : x step size

dy : y step size

rht : fft(rho)

ParticleInCell.SplinePP — Type

SplinePP( degree, ncells)

degree : degree of 1d spline

poly_coeffs : poly_coeffs[i,j] coefficient of $x^{deg+1-j}$ for ith B-spline function size= (degree+1, degree+1)

poly_coeffs_fp : poly_coeffs[i,j] coefficient of $x^{deg+1-j}$ for ith B-spline function size= (degree+1, degree+1)

ncells : number of gridcells

scratch_b : scratch data for b_to_pp-converting

scratch_p : scratch data for b_to_pp-converting

ParticleInCell.SplittingOperator — Type

Operator splitting type for 2d2v Vlasov-Poisson

pic :: PIC poisson solver

pg :: Particle group

ParticleInCell.TwoDGrid — Type

TwoDGrid( dimx, nx, dimy, ny)

Generate a cartesians mesh on rectangle dimxx dimy with nx x ny points

nx : indices are in [1:nx]

ny : indices are in [1:ny]

dimx = xmax - xmin

dimy = ymax - ymin

x, y : node positions

dx, dy : step size